installer: Copy bootloader files from the installer device EFI partition
The installer script (within the installer container) look for EFI files
under /parts/boot. Currently this directory it's not being populated at
all, so the installer script will just create a link for the files from
u-boot container (mounted at /uboot/boot). For default installations there
is no issues on this behavior. However, for the OnLogic FR201 device, the
workflow to produce an installer image is the following:

1. Create the installer raw image
2. Flash the installer image to an USB Stick
3. Change the config.txt file of the EFI partition in the USB Stick to enable
   the options for the FR201 device
4. Boot the USB Stick into the device

For this particular case, the config.txt from the EFI partition of the USB
Stick (installer device) must be copied to the destination EFI partition of
the target storage device (eMMC or SSD). However, the default config.txt
(from the u-boot container) is copied and the device is not able to boot.

This commits adds a function to search for the EFI partition and copy the
files to /parts/boot, so the installer script will install these files
instead of the default ones from u-boot container. In this way, changes in
the config.txt of the installer USB Stick will reflect in the target
device.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>