Remove qemu from installer
Installer installs qemu, but the code that references it in `verify` script is commented out. This means that 5+% of the installer
image, as qemu, is unnecessary. In addition, the booted ubuntu file is not in the image. Removing qemu as it is unused. If we ever
deem it worthwhile to use that space and enable it, it can be added back.

Signed-off-by: Avi Deitcher <avi@deitcher.net>