docs: Add diagram sources for CAN figures
All figures for CAN Bus documentation were produced with draw.io tool.
However, the diagram sources were not embedded to the final PNG file, which
makes impossible to edit the diagrams again using the tool.

This commit replaces all the figures by new ones with the diagram included,
so it can be imported and edited using draw.io.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>