mkimage-iso-efi: Wait for root device inside the initrd script
When installing EVE on devices with Virtual Media CD/DVD-ROM emulation it
can take a few seconds to kernel to detect the installer media device. This
is an issue when booting from an ISO because GRUB will boot kernel +
initrd, but when initrd script tries to find the installer media's
filesystem, the Virtual CD/DVD-ROM can be not yet available. This commit
implements a retry mechanism in order to wait for the root device for up to
10 seconds.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>
(cherry picked from commit e1d09cac9326c63e910c3d1aa6e16c171b3e13a0)