Save EFI BootX variables when (un)sealing storage key
To do a proper analysis a monitor TUI we need values of BootXXXX and
BootOrder EFI variables. Variable data for them is present in the TPM
log however neither TCG spec nor UEFI spec mandate what part of
EFI_LOAD_OPTION structure is serialized and measured in the TPM log so
the reliable extraction is not possible. For example UEFI on NUC13
device avoids first two fields of this structure.

- save variables to /persist/status/boot_vars/{success,fail} directories
  Export paths by introducing GetBootVariablesDirNames()
- export paths to saved TPM logs by introducing GetTpmLogFileNames()
  and GetTpmLogBackupFileNames()

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>