eve: installer: Add rootwait by default
We have experiencing issues with installer due to USB Sticks that take a
few seconds to get detected by the system, leading to "rootfs not found"
errors. This commit adds "rootwait" parameter by default, so the kernel
will wait until root file system is available. The only drawback is that
kernel can wait indefinitely in case of a real failure, but this should be
easily noticed during the installer process. The old behavior can also
be easily configured by overriding dom0_extra_args.

By using rootwait, doesn't make sense to keep the optional entry with
rootdelay option, so it's removed.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>