Makefile: Add platform variant to eve container's TAG and rootfs version
In order to push the eve container for different variants, we must set
different TAGs. Following the same approach for pkg/fw and pkg/nvidia, this
commit adds the platform variant to the eve container's TAG when building
for non-generic platforms. Nothing changes for generic (and specially x86)
platforms.

Additionally, EVE-OS images must be differentiated across platform
variants, so a remote controller can provide the same version for different
platforms, e.g., generic, nvidia-jp5, nvidia-jp6. The platform variant is
also added to the rootfs version (except for the default 'generic'
platform).

The new TAG scheme changes from:

<HASH/EVE_RELEASE>-<HV>-<ARCH>

To:

<HASH/EVE_RELEASE>-<PLATFORM_VARIANT>-<HV>-<ARCH>

The PLATFORM variant is added right after the hash/release in order to
keep the tuple <HV>-<ARCH>, keeping compatibilty with any script/tool
that expects this kind of tag.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>