mkimage-iso-efi: Create a squashfs for the installer rootfs
This commits converts the rootfs of the installer image to a squashfs
image. The initrd script will be responsible to mount this image during
the ISO boot. Using squashfs not only reduces drastically the size of the
final image but also helps to boot ISO when flashed to USB Sticks using
tools like rufus, which will format the USB Stick with a FAT32 partition,
causing several issues when running rootfs from a FAT32 file system.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>