Make mkimage-iso-efi able to build for other architectures
mkimage-iso-efi does everything arch-agnostic, except for if it needs to install
an initrd. In that case, it copies busybox and lib*musl* from its _local_
run, which is whatever architecture it runs on.

This adds the necessary binaries for arm64 and amd64 to all builds from the alpine base,
and then calculates which one to install by the env var ARCH. If ARCH is not set,
it then uses the local arch.

Also updates the README.

Signed-off-by: Avi Deitcher <avi@deitcher.net>