mkimage-iso-efi should always build for architecture of its tar input stream
Previously, mkimage-iso-efi would take the tar as is, but insert initrd for its local platform, which made no sense.
An earlier commit fixed it so you could specify TARGETARCH. Thus, if running on am64 and your tar is arm64, you could force it
to build for arm64 via TARGETARCH=arm64. That works, but even that is too complex.

There is no reason *ever* to build the initrd for anything other than the architecture _of the input tar stream_.
Rather than trying to use local arch, or have the user signal it with an env var, this just takes the arch of the tar stream and uses it.

mkimage-iso-efi always will build the initrd for the architecture of the input stream. Much cleaner and simpler.

Kudos to Rene Pinto, who thought of this nice one.

Signed-off-by: Avi Deitcher <avi@deitcher.net>