pillar: Add passed-through devices to cgroups device list as well
When passingthrough a device to a Reduced Isolated container, the file
device is added only to the device list in the OCI spec. However, in order
to allow full access to the device, it must also be added to the cgroups
device list.

The error was observed while passingthrough a webcam to a Reduced Isolated
container. The gst-launch application returned the following error:

Setting pipeline to PAUSED ...
ERROR: from element /GstPipeline:pipeline0/GstV4l2Src:v4l2src0: Could not open device '/dev/video0' for reading and writing.
Additional debug info:
../sys/v4l2/v4l2_calls.c(627): gst_v4l2_open (): /GstPipeline:pipeline0/GstV4l2Src:v4l2src0:
system error: Operation not permitted
ERROR: pipeline doesn't want to preroll.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>