Fix flow-logging config removal
When the last NI with flow-logging enabled is being deleted,
LinuxNIReconciler should remove all config items used solely for the
flow logging purposes (e.g. the "blackhole" dummy interface).
However, due to a bug in the withFlowlog method this was not the case.
The method would return true even when the remaining non-deleted NIs
had all flow logging disabled.

Signed-off-by: Milan Lenco <milan@zededa.com>