Remove arch-specific extension to package names when building rootfs.yml
A long time ago, we published distinct images for different architectures, so you had to refer to them by specific name
in order to get the right image. Further, the underlying tooling didn't support resolving the correct architecture.

That is a long-time past.

First, all of our images are OCI indexes, so we do not need to specify `lfedge/eve-pillar:abcd-amd64`, but can just do
`lfedge/eve-pillar:abcd`. Second, our tooling properly resolves the correct underlying image for the target architecture
from the OCI index.

Finally, using direct images, rather than indexes, means we lose access to other information stored on the index about the image,
but not in the image itself. The obvious example is sbom, but other attestations and data are stored there as well.

Signed-off-by: Avi Deitcher <avi@deitcher.net>