Publish info and metrics also for L2-only ports
It was decided that it makes sense to publish info and metrics also for
ports which do not have Network assigned (without SystemAdapter in
EdgeDevConfig) but still are used by VLANs or LAGs. For example,
user may configure "vlan-10" and "vlan-20" for "eth0", use one of the
VLAN sub-interfaces for management, the other for apps, while not using
eth0 directly with any network or network instance. Still it makes sense
for EVE to collect and publish metrics and status information for the VLAN
parent interface.
Note that NICs which are not used directly, only through VLANs/LAGs,
still have NetworkPortStatus entry inside DeviceNetworkStatus, but IsL3Port
attribute is set to false.

Signed-off-by: Milan Lenco <milan@zededa.com>