msrv: Add /networks/metrics.json endpoint
On some use cases an Edge App can be used to check the health of the
Edge Node, requiring to receive the metrics and status of all device
ports (network interfaces). This commit adds the /networks/metrics.json
endpoint to the meta data server so VMs can use it to retrieve the
corresponding information. The information is provided through a JSON
output, like the following:

[{
    "IfName": "eth0",
    "Up": true,
    "TxBytes": 405385,
    "RxBytes": 3375652,
    "TxDrops": 0,
    "RxDrops": 0,
    "TxPkts": 2962,
    "RxPkts": 5739,
    "TxErrors": 0,
    "RxErrors": 0,
    "TxACLDrops": 0,
    "RxACLDrops": 0,
    "TxACLRateLimitDrops": 0,
    "RxACLRateLimitDrops": 0
}]

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>