alpine-base: Bump minirootfs to version 3.16.9
This is the first step to bump current Alpine 3.16.2 to 3.16.9 so we can update
packages and get a newer Go version.

This commits bumps the minitroofs version for x86_64 and arm64 to 3.16.9 and
busybox version from 1.34.1 to 1.35.0, used by Alpine 3.16.9.

PS1: Changes are valid only for amd64 and arm64. The riscv64 is left
untouched for now.

PS2: alpine 3.21 image is used as builder due to a bug with busybox that
makes wget to fail when downloading images for arm64 and riscv64 running
from an emulated environment (on docker).

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>