alpine: Make a new bootstrap to update to 3.16.9
This is the second step towards the bump of Alpine to 3.16.9. Package
pkg/alpine-base was built with the minirootfs of 3.16.9 version.

This commit uses the latest pkg/alpine-base as the base image and
provides the changes to build pkg/alpine bumping it to 3.16.9. Once this
package is published, it can be used as the base image for itself
(pkg/alpine).

PS: Changes are only applied to amd64 and arm64. Another base image
(without changes) is used for riscv64.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>
(cherry picked from commit 458c9ebc799e9d36beddf18e76a4459c99a7e33f)
Signed-off-by: Paul Gaiduk <paulg@zededa.com>