newlog: add filter, dedup and counter functions
Newlog will have 3 ways to reduce the amount of logs:
- filter: filter logs out based on the source code line that
produced them
- counter: count the number of logs produced by a specific
source code line. Add that number to the first occurance of the log
and remove the rest
- deduplicator (for errors only): record the last X errors in a sliding
window and remove duplicates

The benchmarking of the newlogd with the new features is in the
dedup_test.go file. It shows that CPU and RAM usage increase by a factor
of 3 when the features are enabled. So they can be disabled by setting
the deduplication window size to 0 and not providing anything to the
filter and counter functions.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>