volumemgr: recreate CT on significant changes
When the content tree config parameters are changed (but the UUID stays
the same) we need to differentiate between significant and
insignificant changes. The significant changes include the change to the
content itself, like change of the URL or the checksum.
When such changes are detected, we need to remove the old content tree
including the old blobs and create a new one with the new content.
Otherwise the changes will not take effect.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>