Add qmi-firmware-update to enable modem firmware upgrades from EVE
The qmi-firmware-update utility allows firmware upgrades for cellular
modems over the QMI interface. Including it in the wwan container
enables users to upgrade modem firmware directly from EVE, eliminating
the need to install another OS or passthrough the modem to a VM with
a separate upgrade tool.

Note that the binary was already being built with libqmi, but it wasn’t
included in the final wwan container image.

The increase in EVE image size is minimal, as the binary is only 226KB.

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit dd7924d7ef0ff497e12678c0bfed4edbe4ffb6d9)