vTPM backup swtpm state and check health
SWTPM state is backed up before every write and vTPM first checks the
health of the SWTPM before handing it over to the domain manager.

Based on the result of the health check, either normal state is used or
the backup state is used or if both are corrupted/missing, vTPM
will reset the SWTPM state.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>