Handle WirelessType NOOP
When a port is not a wireless network adapter, the controller may either
return nil for the WirelessConfig or provide a WirelessConfig with Type
set to NOOP. These two cases are functionally equivalent.
However, EVE returns an "unsupported wireless type" error when encountering
the NOOP type.

Previously, this wasn’t an issue because errors from parseNetworkWirelessConfig
were simply logged and ignored.
After recent refactoring of this function, those errors are no longer ignored
and will cause EVE to reject the entire network configuration as invalid.

Signed-off-by: Milan Lenco <milan@zededa.com>