Set parent interface UP before bringing up VLAN subinterface
Ensure the parent interface is set to UP before bringing up
a VLAN subinterface. Otherwise, netlink.LinkSetUp(vlan) returns
a "network is down" error.

This issue hasn’t surfaced until now because in the performed VLAN tests:
- the parent ethernet interface was also used for untagged traffic and
  was therefore coincidently brought UP beforehand, or
- the device was bootstrapped with a last-resort config, which brings
  all Ethernet interfaces UP.

However, when bootstrapping a device using a config that includes VLANs
but does not use the parent interface directly (i.e. it is in L2-only
mode), the VLAN setup fails with the "network is down" error because
we forgot to bring the parent UP.

Signed-off-by: Milan Lenco <milan@zededa.com>