be sure to include config.img/grub.cfg override in installer
The installer needs to check for a `config.img` in its installer, and if it finds one,
look for `grub.cfg` in it. If it finds it, it should source it. It did not, this now adds it.

Signed-off-by: Avi Deitcher <avi@deitcher.net>
(cherry picked from commit 7ff87934f5294aefef315bfba33dc6bedf8a200d)