Fix broken run-* targets for dirty builds
All run-* targets must not build any artifacts because it causes an update
of ./dist/<arch>/current symbolic link and thus run-* target cannot find artifacts
anymore. Instead all references to artifacts in run-* targets must be relative
to 'current'. It was not a problem if changes were committed because new build directory
did not contain current time but in case of dirty builds new directory was created every time

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>