Remove TPM eventlog parsing/sending in attestation process
Remove the TPM eventlog parsing and sending in the attestation process,
the current implementation ommits the binary data for request size
reduction and this basically makes it impossible to use the eventlog
for for validation or implementing policies. TPM event log should be
sent as a binary blob in the attestation request, so that it can be
used for validation in the future.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>