Introduce Air-gap mode
Air-gap mode was added to support scenarios where a device operates
without physical connectivity to the main controller. In such cases,
a Local Operator Console (LOC) is used to manage the device locally.

When air-gap mode is enabled and a LOC URL is configured,
the connectivity tester will report the device as "connected" if either
the main controller or the LOC is reachable. This ensures that when
a user makes a network configuration change, EVE does not flag
the configuration as non-working—and does not revert to the previous
configuration—as long as LOC connectivity is maintained.

Additionally, when in air-gap mode, if the device fails to access
the main controller, error and warning logs related to the lost
connection are completely suppressed or at least downgraded to the trace
level. This prevents log clutter, as lack of controller connectivity
is expected in this mode. However, EVE will continue attempting
to communicate with the main controller in case connectivity is restored
(at least temporarily).

Signed-off-by: Milan Lenco <milan@zededa.com>