Fixes VirtualBox EVE VM onboarding
Ensures VirtualBox EVE VMs can onboard correctly by setting the SMBIOS serial number.

Also includes minor fixes to network configuration for the VMs.
Moreover, added the current makefile target as a dependency to live-% so when we run ```make live-vdi``` the current symlink will point to the latest build that contains the vdi image. When then we run run-live-vb we use the current symlink directory to create the vbox VM

Signed-off-by: Othonas Soultatos <othonas.s@gmail.com>