Remove EVE_REL from tags
We had a strange logic, where if the HEAD commit has a tag on it, *and* it is the highest
semver tag available, then instead of building and pushing packages with the usual
tree hash and also adding the tag, we actually override the usual and force the tag.
This is strange and bad in two ways.

First, it means that even if nothing changes, it forced a rebuild of every single package. Which is silly. We want to use the existing ones, and just add the tag.
Second, it means that we did it only for the most recent semver. So if we already have 15.2.0 and 14.5.0, then 15.3.0 would trigger this behaviour, but 14.5.1 would not? That makes no sense.

This eliminates it. From this point forward, every tag that is on HEAD gets --release to add the tag.

Signed-off-by: Avi Deitcher <avi@deitcher.net>
(cherry picked from commit 3f291cf39ef474a8f30c659c284a8cf7d6e47499)
Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>