Include device uuid to the filename of collectinfo archive
When we are using collect-info.sh the compressed file name does not contain any information about the device and it can be hard to distinguish the origination of the file if you want to generate info for more than one device. This commit changes the script to get the UUID of the device if it is onboarded, if not, it used the device serial.

Signed-off-by: Othonas Soultatos <othonas.s@gmail.com>