Introduce /etc/eve-platform file
A lot of scripts e.g. installer, make-raw must know which platform we are
running on but it is very hard to pass it at runtime. It would require us to have
YQ modifiers for several eve packages JUST to pass a buildArg. Instead we read it
from /etc/eve-platform

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>