New image layout with IMGC support
'make-raw' is a backbone script that write rootfs images to disk and controls
partition table layout. It may be called in 3 different scenarios:
1. From Makefile as a container using Docker
2. As a script from installer (it is mounted as volume)
3. From docker run --rm lf-edge:eve <target>

In [2,3] the platform is read from /etc/eve-platform, in [1] it is passed
from docker command line as -e PLATFORM=. Besides the list of required
partitions can be omitted when calling the script

This commit introduces IMGC together with its GUID and functions
to handle IMGB and IMGC and calculate their size. Since we never update
evaluation image as we do for regualar EVE we do not need to reserve space
in neither partitions to minimize installer size

Also
- remove unused code
- fix some shell syntax

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>