pillar/zedagent: Check BaseOS Content Tree
There are some corner cases where users push an EVE OS image for update but
it can cancel the operation while the download didn't start or it's not
finished yet. This commit adds some checks to validate the content tree
before report the OS Image as valid to the remote controller.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>