dom0-ztools: Replace pre-cooked UEFI settings
The pre-cooked UEFI settings had a boot order with the following
indexes:

0001, 0005, 0002, 0003, 0000

In some specific configurations, specially on Edge Apps with multiples
disks, this can lead to boot to the EFI Shell. This commit fixes the boot
order to the indexes:

0000, 0001, 0002, 0003, 0004, 0005

Additionally, a new settings file is added with 640x480 default video
resolution.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>