pkg/bpftrace: disable tests when cross-compiling
1. running the tests with qemu-bin emulation is slow
2. the tests might actually not work because of the different
   architecture

When running the bpftrace tests inside the container, we want to skip
some or all tests if the environment is likely to be slow or unreliable,
such as when using QEMU user-mode emulation (commonly seen when building
on Mac hosts for x86, especially Apple Silicon).

Since /proc/cpuinfo inside Docker on Mac does not clearly indicate the
host platform, we use a heuristic: check the "bogomips" value for each
CPU. Bogomips is a rough measure of processor speed as reported by the
Linux kernel; when running under emulation, this value is often very low.

We use a threshold of 500 bogomips, which is low enough to catch Macs,
but high enough to avoid false positives on other
hardware (at least in my tests).
If any CPU reports less than 500, we assume we are running in
an emulated or virtualized environment on a MAC and skip some tests that are known
to be unreliable or too slow there.

Also, if TARGETARCH and BUILDARCH do not match, all tests are skipped,
as cross-arch emulation is expected.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>