pillar: Do not rename package for kubevirt variants
Instead of use a different package name for kubevirt variants, such as:

- eve-pillar-kube
- eve-pillar-kube-dev

Use the ability of linuxkit to set a custom tag from the build.yml file.
Thus, both variants will have the same eve-pillar image name, but with
different hashes. For instance:

- eve-pillar:a85a6e51b4eb0cde41e0e43ae2d0f9e58b30ea06-kube
- eve-pillar:a85a6e51b4eb0cde41e0e43ae2d0f9e58b30ea06-kube-dev

The same scheme is also applied to the eve-pillar-dev image.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>