Port Makefile changes to improve/fix linuxkit build
This commit ports (adapted) the following commits:

Upgrade Linuxkit to 1.7.0

This version has many important fixes
1. Import of Docker credentials
2. --builder-config can pass a buildkitd.toml to a builder container

So we do not need an ensure-builder target anymore. It was very useful
for debugging Docker login / mirror related issues but I hope it is gone
for good.

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>

Create linuxkit-builder manually to support parallel build

Linuxkit tries to create docker builder container if it doesn't exist.
It works fine for -j 1 but fails for parallel build. LK tries to create or update the
builder concurrently and fails. To avoid a race we create builder manually

- the builder name is hard-coded and is linuxkit-builder
- override context name for `linuxkit pkg build` so it doesn't try
to create a builder by itself
- ensure-builder target checks if the container exists and of required version
- docker prune operation cannot be executed concurrently, use flock to handle that
this is especially important for CI build with PRUNE=1 set

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>

Fix 'unauthenticated pull rate limit reached' problem

- action/docker-login stores credential in ~/.docker/config.json. In theory this file should be mounted
into linuxkit-builder container but I did not know about it. Dut to that fact buildkit was running unauthenticated
- upgrade to linuxkit 1.6.5 that has fixes for a race in builder creation and cache access
- mount /etc/buildkit/buildkitd.toml that MAY define a mirror to speedup the build. LK doens't do it
other wise we could remove the whole block for manual creation of buildkit runner

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>

Makefile: Bump linuxkit to version 1.5.3

Bump linuxkit to version 1.5.3 in order to get a new feature that supports
--tag in build.yml. This version also introduces some changes that
requires some adjusts in the Makefile:

1. Introduces different verbose levels (this commit sets to level 2 to keep
   the same amount of information provided by the current version)
2. Remove --arch parameter from cache command (in favor of a --platform
   docker style parameter)

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>