Create linuxkit-builder manually to support parallel build
Linuxkit tries to create docker builder container if it doesn't exist.
It works fine for -j 1 but fails for parallel build. LK tries to create or update the
builder concurrently and fails. To avoid a race we create builder manually

- the builder name is hard-coded and is linuxkit-builder
- override context name for `linuxkit pkg build` so it doesn't try
to create a builder by itself
- ensure-builder target checks if the container exists and of required version
- docker prune operation cannot be executed concurrently, use flock to handle that
this is especially important for CI build with PRUNE=1 set

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>