.github: Partially roll back tag normalization in assets workflow.
Restore use of ${{ inputs.tag_ref }} directly in the release creation step,
removing the Bash variable and tag normalization logic. This is necessary because
Bash variables are not expanded inside single-quoted strings, causing the literal
"${TAG}" to be sent to the GitHub API instead of the intended tag name.

Other unrelated changes from the previous commit are retained.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>
Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>