.github: Remove kubevirt from PR build matrix
Temporarily disable the kubevirt variant in the PR Build workflow. This
follows up on a previous change by removing the last reference to
kubevirt, which was inadvertently left in the matrix.

The kubevirt build is currently failing and blocks CI testing. Disabling
it allows development to proceed while the issue is investigated.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>