Avoid endless loop in DPCManager.runVerify when we get RemoteTemporaryFailure
This commit fixes a bug in DpcManager.runVerify that would cause an infinite
loop when a RemoteTemporaryFailure was returned by the ConnectivityTester.

The issue was that DPCManager didn’t flag the DPC as tested when a RemoteTemporaryFailure
was received. Such a DPC was then mistakenly evaluated as a new DPC received during
testing, and the entire testing procedure was repeated—resulting in an endless cycle.

This commit fixes the issue by flagging the DPC as tested, and actually as
successful, reporting the RemoteTemporaryFailure only as a warning,
since connectivity itself is working.

There is no need to handle testing retries for RemoteTemporaryFailure explicitly,
as they are already performed periodically every 5 minutes (timer.port.testinterval),
even when the latest DPC is marked as successful.

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit 07215ab97b6e31cb1d89019066e8f8bf0755d7fa)