Do not prune builtkit builder for parallel build
For parallel build buildkit builder may be used by other targets,
we must not remove it. Besides it is reused but all other build
The only downside is that the container stays if we move to newer
version. However commit

e0726b6 - Create linuxkit-builder manually to support parallel build

fixes this problem.

Also update linuxkit to 1.6.4 that has important fixes for cache locking

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>