Handle error IsAlreadyExists
During a failover scenario, kubernetes starts the VMI on the other node, because we just use kubernetes
scheduling mechanism. Then we tell eve domainmgr to process the VMI start and publish the DomainStatus.
So when domainmgr calls the Start() we need to handle the already exists error and return gracefully.
Without this fix after failover domainmgr code goes into forever loop.

This code is present in POC branch, we missed in merges.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>