Handle isreplicated flag for the volumes
A volume in an edge cluster has a designated owner either its created through an app or explicitly.
We distinguish those using the flag isreplicated.
isreplicated = false  (owner node of the voulme)
isreplicated = true (volume is getting replicated to this node)

1) Do not delete the PVC if is a replicated volume.
2) Do not do volume GC if it is a replicated volume.
3) Ignore volume process in doUpdateVol() if it is replicated volume.
4) CreateVolume() for replicated volumes will just wait for PVC to be created by owner node.

We defined the isreplicated flag in previous commits, but missed the merges of handling it.
This commit brings in the code to handle the isreplicated flag from POC branch.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>