Cleanup Replica sets only in single node mode
This commit fixes a serious bug in cluster configuration.
Basically whenever domainmgr restarts there was a legacy code which deletes all
VMs running. That was implemented to handle a corner case where apps were deleted when device is
powered off. We want to delete any app config in kubernetes as soon as device reboots. Assumption is
device will resync config with the controller.

That works fine for single node cases. In a cluster setup, apps would have moved to other nodes when this
device went to reboot. Because of this bug as soon as this device reboots its deleting VMs running on other nodes,
basically bouncing all VMs.

Added a new API to check if this device is in cluster mode and ignore delete of VMs.

Also, we no longer directly deal with VMs, its all VM replica sets now. So moved CleanupVMs to CleanupVMRs.

Move the const to types/locationconstants.go
Change the logging to Debugf

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>