Return unknown error if k3s is not reachable
If we cannot reach kubernetes to get the status of VM, handle that error and return
"unknown" status for domainmgr. This could happen if k3s is down or network cable was pulled
on the device to remove the device from the cluster.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>
(cherry picked from commit 2bac6de21578115ac2de9d290dd3bde5d023bdfa)