Handle apps during explicit node reboot
When the user triggers a reboot of a device from Controller,
domainmgr tries to do DomainShutdown (app deactivate), waitforDomain gone and then
in worst case Deletes domain.

That is all good and expected procedure on a single device configs.
In a cluster mode the apps would have failed over to another node and we cannot
delete the app. So handle such cases.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>